Refactored code to work with new launcher thread group. Working much better now. Removed some code, saved in _here_lays_old_code_.py may need this in the future.
Changed the Environment class to a dataclass. Moved the gitdeploy package to the root of the flask app. This imports better.

The launching of the flask app and supervisor are now done in start.py, the launching of supervisorctl is done by the flask app. This gives better control to the flask app.

Some code cleanup was done as well, some functions in there that were not getting used.